+++ subscriber/podaac_data_subscriber.py
@@ -14,17 +14,16 @@
 # Accounts are free to create and take just a moment to set up.
 import argparse
 import logging
-import os, re
+import os
 import sys
 from datetime import datetime, timedelta
 from os import makedirs
 from os.path import isdir, basename, join, isfile, exists
 from urllib.error import HTTPError
-from urllib.request import urlretrieve
 
 from subscriber import podaac_access as pa
+from subscriber import subsetting
 from subscriber import token_formatter
-
 
 __version__ = pa.__version__
 
@@ -107,6 +106,7 @@ def create_parser():
     parser.add_argument("-p", "--provider", dest="provider", default='POCLOUD',
                         help="Specify a provider for collection search. Default is POCLOUD.")  # noqa E501
     parser.add_argument("--dry-run", dest="dry_run", action="store_true", help="Search and identify files to download, but do not actually download them")  # noqa E501
+    parser.add_argument("--subset", dest="subset", action="store_true", help="Subset the data via Harmony calls.")  # noqa E501
 
     return parser
 
@@ -236,20 +236,85 @@ def run(args=None):
         logging.info(str(results[
                              'hits']) + " new granules found for " + short_name + " since " + data_within_last_timestamp)  # noqa E501
 
+    file_start_times = None
+    cycles = None
     if any([args.dy, args.dydoy, args.dymd]):
         file_start_times = pa.parse_start_times(results)
     elif args.cycle:
         cycles = pa.parse_cycles(results)
 
-    timestamp = datetime.utcnow().strftime("%Y-%m-%dT%H:%M:%SZ")
+    granules = results['items']
+    granules = list(map(lambda granule: granule['meta']['concept-id'], granules))
 
+    collection_id = pa.get_cmr_collection_id(
+        collection_short_name=args.collection,
+        provider=args.provider,
+        token=token,
+        verbose=args.verbose
+    )
+
+    subsettable = False
+    if args.subset:
+        subsettable = subsetting.is_subsettable(
+            collection_id=collection_id,
+            token=token,
+        )
+
+    if subsettable:
+        success_cnt, failure_cnt = subsetting.subset(
+            collection_id=collection_id,
+            start_date_time=args.startDate,
+            end_date_time=args.endDate,
+            bbox=args.bbox,
+            force=args.force,
+            data_path=data_path,
+            args=args,
+            process_cmd=process_cmd,
+            granules=granules
+        )
+    else:
+        success_cnt, failure_cnt, skip_cnt = cmr_downloader(
+            granules=results['items'],
+            extensions=extensions,
+            args=args,
+            data_path=data_path,
+            file_start_times=file_start_times,
+            ts_shift=ts_shift,
+            cycles=cycles,
+            process_cmd=process_cmd
+        )
+
+    if len(granules) > 0 and not args.dry_run:
+        if not failure_cnt > 0:
+            with open(f'{data_path}/.update__{args.collection}', 'w') as f:
+                f.write(datetime.utcnow().strftime("%Y-%m-%dT%H:%M:%SZ"))
+
+    if success_cnt > 0:
+        try:
+            pa.create_citation_file(short_name, provider, data_path, token, args.verbose)
+        except:
+            logging.debug('Error generating citation', exc_info=True)
+
+    logging.info('END\n\n')
+
+
+def cmr_downloader(granules, extensions, args, data_path, file_start_times, ts_shift, cycles, process_cmd):
     downloads_all = []
-    downloads_data = [[u['URL'] for u in r['umm']['RelatedUrls'] if
-                       u['Type'] == "GET DATA" and ('Subtype' not in u or u['Subtype'] != "OPENDAP DATA")] for r in
-                      results['items']]
-    downloads_metadata = [[u['URL'] for u in r['umm']['RelatedUrls'] if u['Type'] == "EXTENDED METADATA"] for r in
-                          results['items']]
-    checksums = pa.extract_checksums(results)
+
+    downloads_data = [
+        [
+            u['URL'] for u in r['umm']['RelatedUrls']
+            if u['Type'] == "GET DATA"
+               and ('Subtype' not in u or u['Subtype'] != 'OPENDAP DATA')
+        ] for r in granules
+    ]
+    downloads_metadata = [
+        [
+            u['URL'] for u in r['umm']['RelatedUrls']
+            if u['Type'] == 'EXTENDED METADATA'
+        ] for r in granules
+    ]
+    checksums = pa.extract_checksums(granules)
 
     for f in downloads_data:
         downloads_all.append(f)
@@ -281,8 +346,7 @@ def run(args=None):
         for download in downloads:
             logging.info(download)
         logging.info("Dry-run option specific. Exiting.")
-        return
-
+        return 0, 0, 0
 
     # NEED TO REFACTOR THIS, A LOT OF STUFF in here
     # Finish by downloading the files to the data directory in a loop.
@@ -302,14 +366,13 @@ def run(args=None):
                     cycles, data_path, f)
 
             # decide if we should actually download this file (e.g. we may already have the latest version)
-            if(exists(output_path) and not args.force and pa.checksum_does_match(output_path, checksums)):
+            if (exists(output_path) and not args.force and pa.checksum_does_match(
+                    output_path, checksums)):
                 logging.info(str(datetime.now()) + " SKIPPED: " + f)
                 skip_cnt += 1
                 continue
 
-            #urlretrieve(f, output_path)
-            pa.download_file(f,output_path)
-
+            pa.download_file(f, output_path)
             pa.process_file(process_cmd, output_path, args)
             logging.info(str(datetime.now()) + " SUCCESS: " + f)
             success_cnt = success_cnt + 1
@@ -317,27 +380,11 @@ def run(args=None):
             logging.warning(str(datetime.now()) + " FAILURE: " + f, exc_info=True)
             failure_cnt = failure_cnt + 1
 
-    # If there were updates to the local time series during this run and no
-    # exceptions were raised during the download loop, then overwrite the
-    #  timestamp file that tracks updates to the data folder
-    #   (`resources/nrt/.update`):
-    if len(results['items']) > 0:
-        if not failure_cnt > 0:
-            with open(data_path + "/.update__" + short_name, "w") as f:
-                f.write(timestamp)
-
     logging.info("Downloaded Files: " + str(success_cnt))
     logging.info("Failed Files:     " + str(failure_cnt))
     logging.info("Skipped Files:    " + str(skip_cnt))
 
-    if success_cnt > 0:
-        try:
-            pa.create_citation_file(short_name, provider, data_path, token, args.verbose)
-        except:
-            logging.debug("Error generating citation", exc_info=True)
-
-    logging.info("END\n\n")
-
+    return success_cnt, failure_cnt, skip_cnt
 
 def main():
     log_format = '[%(asctime)s] {%(filename)s:%(lineno)d} %(levelname)s - %(message)s'